split CMDHelperFuncs from otaclient.app.boot_control.common as a new module cmdhelper
